\name{horror_movies}
\alias{horror_movies}
\docType{data}
\title{
Horror Movies Metadata
}
\description{
}
\usage{data("horror_movies")}
\format{
  A data frame with 3,328 observations on the following 17 variables.
  \describe{
    \item{\code{release_country}}{Release country}
    \item{\code{movie_rating}}{MPAA Rating}
    \item{\code{review_rating}}{Movie rating (0 - 10)}
    \item{\code{movie_run_time}}{Movie run time (minutes)}
    \item{\code{language}}{Language}
    \item{\code{budget}}{Budget (US Dollars) log10 transformed}
    \item{\code{action}}{Is it an action movie?}
    \item{\code{adventure}}{Is it an adventure movie?}
    \item{\code{comedy}}{Is it a comedy movie?}
    \item{\code{crime}}{Is it a crime movie?}
    \item{\code{drama}}{Is it a drama movie?}
    \item{\code{fantasy}}{Is it a fantasy movie?}
    \item{\code{mystery}}{Is it a mystery movie?}
    \item{\code{romance}}{Is it a romance movie?}
    \item{\code{sciFi}}{Is it a sci-fi movie?}
    \item{\code{thriller}}{Is it a thriller movie?}
    \item{\code{release_year}}{Movie release year}
  }
}
\details{
}
\source{
https://github.com/rfordatascience/tidytuesday/tree/master/data/2019/2019-10-22
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(horror_movies)
}
\keyword{datasets}
